Make cli work with runtime
