+++ cli/package.json
@@ -26,14 +26,14 @@
         "winston": "catalog:"
     },
     "devDependencies": {
-        "@tsconfig/node16": "catalog:",
-        "tslib": "catalog:",
-        "tsx": "catalog:",
-        "typescript": "catalog:",
+        "@tsconfig/node22": "catalog:",
+        "@types/blessed": "^0.1.25",
+        "@types/copy-paste": "^1.1.33",
         "prettier": "catalog:",
         "prettier-plugin-organize-imports": "catalog:",
-        "@types/blessed": "^0.1.25",
-        "@types/copy-paste": "^1.1.33"
+        "tslib": "catalog:",
+        "tsx": "catalog:",
+        "typescript": "catalog:"
     },
     "files": [
         "dist"
