+++ cli/tsconfig.json
@@ -1,12 +1,13 @@
 {
-    "extends": "@tsconfig/node16/tsconfig.json",
+    "extends": "@tsconfig/node22/tsconfig.json",
     "compilerOptions": {
         "downlevelIteration": true,
         "incremental": true,
         "removeComments": true,
         "rootDir": ".",
         "declaration": false,
-        "outDir": "dist"
-      },
-      "exclude": ["specs/", "dist/"]
+        "outDir": "dist",
+        "resolveJsonModule": true
+    },
+    "exclude": ["specs/", "dist/"]
 }
